new directories and changes to servlets
This commit is for the new folders ‘student’ and ‘admin’ in the Web
Content folder. I changed the path in the servlets to match this
change. I added ‘/student/index.html’ and ‘’/student/index.jsp’ in the
web.xml file.

I didn’t commit the .classpath because it contains the tomcat server to
use.