user login, user class, and user view reservations
changed LoginController and Login Servlet to get the user who
authenticated (hard coded for now), check to see if they are in the
user table. If so, update their last login field, and if not add them
to the user table.

created/modified user UserHelper to do the dirty work for the login
controller, login servlet, and user view servlet.

changed web folder ‘student’ to ‘user’ and modified code to accept this
change.

added more data to display on the jsp’s for testing.

added (or modified) user class to be used instead of student class

changed the web.xml file to point to the user folder and not the
student folder