added banned check to login
I added the check to see if they have already been banned for this
period. If they have not been banned they will be directed to the home
page. If they have been banned, they will be sent to a banned page with
information and a button to log out.