Updates to Admin Reserve + Time Range Method
— changed time range method to a string and it now prints out the time
in 24 hour format instead of an hour integer
— removed some line breaks
— admin reservations will tell you if a room is reserved or if its
free.  If free, a button that allows your to make a reservation will
appear beside the room number, etc.  If reserved, there will be no
button but the text “**reserved**” in its place.
— created a method and query in Reservation Select Query that accepts
the end time.  The query will check the start date for a reservation as
well as the start date - end date for a reservation.
— admins can only make a reservation between 2 days.  Doing more is too
complicated.
— ADMIN RESERVE AND ADMIN SCHEDULE ADD ARE NOT COMPLETE YET!!