New Time methods + Fixing issues.
— TimeConverter.addOneSecondToTime is self explanatory and takes an
inputted time in sql format
— TimeConverter.subtractOneSecondToTime is self explanatory and takes
an inputted time in sql format
— These methods address the issue of querying reservations.  The end
time should *not* be the same as the start time for the next
reservation.  This solves the issue of the *between* sql operator in
which the values are inclusive.
— used the new methods for reservationEndTime (inserting into the db or
displaying
— fixed building select drop down to display a message if building is
currently closed
— added a method to get the current minutes.  This is used in browse
reservation.  If room isn’t reserved but is 10min past hour, then room
can’t be reserved and is highlighted yellow to show that the room is
open/free for all.
— will focus on schedule add and admin reservation.