CAS testing changes, figuring out edit rooms link from building list
CASLoginServlet - Changed println’s, added check to see if the test1010
user logs in, the employee type is 01, but I wanted to test the admin
side, so I put an if statement. Feel free to comment out.

BuildingListQuery - changes to code to see if can figure out the
displaying/linking issue. Changes minor.

ListUserReservationsQuery - added test lines for checking in too late

ConfirmCancellation - change display

web.xml: changed it back to original - still get a red square but I see
no difference… must be a space or a tab